Adds smartcard attachment tracking
Adds HardwareThread.
We are now auto-selecting the top row in the list.

All dialogs will stop the HardwareThread object of the mainwindow,
and restarts it when dialog is done or closed.

The exit event or the action will take a second, it will wait
for the HardwareThread to stop first.